+++ jest.config.js
@@ -250,7 +250,7 @@ module.exports = {
 
   // An array of regexp patterns that are matched against all source file paths before re-running tests in watch mode
 
-  // watchPathIgnorePatterns: [],
+  watchPathIgnorePatterns: ['stories/*'],
 
   // Whether to use watchman for file crawling